refactor(agents): split trace record logic
